Refactor Leaflet components
